# contracts/openapi/watchman.v1.yaml
openapi: 3.1.0
info:
  title: Atalaya Watchman API
  version: 1.0.0
  description: API privada consumida por el Command Center. La sesión viaja en una cookie HttpOnly.
servers:
  - url: /api/v1
paths:
  /session:
    post:
      summary: Iniciar sesión de usuario único
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties: { password: { type: string, format: password } }
      responses:
        '200': { description: Sesión creada; devuelve cookie HttpOnly }
        '401': { description: Credenciales inválidas }
    get:
      summary: Validar la sesión actual
      responses:
        '200': { description: Sesión válida }
        '401': { description: Sin sesión }
    delete:
      summary: Revocar la sesión actual
      responses:
        '204': { description: Sesión revocada }
  /overview:
    get:
      summary: Obtener estado general, integraciones y eventos recientes
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Overview operativo } }
  /events:
    get:
      summary: Listar eventos con filtros y paginación
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: application, in: query, schema: { type: string } }
        - { name: severity, in: query, schema: { enum: [critical, high, medium, low, pending] } }
        - { name: state, in: query, schema: { enum: [actionable, noise, pending] } }
        - { name: period, in: query, schema: { type: string, examples: [24h, 168h] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses: { '200': { description: Página de eventos } }
  /events/{id}:
    get:
      summary: Obtener evento, interpretación y detalle técnico
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Detalle del evento }
        '404': { description: Evento inexistente }
  /integrations:
    get:
      summary: Obtener salud de fuentes y checkpoints
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Integraciones operativas } }
  /reports:
    get:
      summary: Listar el historial de reportes diarios
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 30 } }
      responses:
        '200':
          description: Reportes diarios con la fotografía de cada aplicación
  /error-groups:
    get:
      summary: Buscar grupos de error para asociarlos a incidentes
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: application, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses: { '200': { description: Grupos ordenados por última aparición } }
  /incidents:
    get:
      summary: Listar incidentes
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Historial de incidentes } }
    post:
      summary: Crear un incidente en investigación
      security: [{ cookieAuth: [] }]
      responses: { '201': { description: Incidente creado }, '409': { description: Un grupo ya está bajo investigación } }
  /incidents/{id}:
    get:
      summary: Obtener grupos y bitácora inmutable del incidente
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Detalle del incidente } }
  /incidents/{id}/notes:
    post:
      summary: Agregar una nota al incidente
      security: [{ cookieAuth: [] }]
      responses: { '201': { description: Nota agregada } }
  /incidents/{id}/status:
    patch:
      summary: Resolver, marcar como ruido o reabrir con una conclusión obligatoria
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Estado actualizado }, '409': { description: Reapertura incompatible } }
  /incidents/{id}/publication:
    patch:
      summary: Publicar, editar o retirar la descripción pública de un incidente
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [published]
              properties:
                published: { type: boolean }
                title: { type: string, maxLength: 160 }
                message: { type: string, maxLength: 2000 }
      responses: { '200': { description: Publicación actualizada }, '400': { description: Contenido inválido } }
  /public/status:
    get:
      summary: Estado público de aplicaciones e incidentes publicados
      security: []
      responses:
        '200':
          description: Estado público sin datos técnicos
          headers:
            Cache-Control: { schema: { type: string } }
  /system/health:
    get:
      summary: Heartbeats, pollers y métricas de colas internas
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Salud interna detallada } }
  /deployments:
    get:
      summary: Listar deploy markers
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Deploys exitosos de producción } }
    post:
      summary: Registrar un deploy manual exitoso
      security: [{ cookieAuth: [] }]
      responses: { '201': { description: Deployment creado } }
  /operations/timeline:
    get:
      summary: Obtener buckets de errores, deploys e incidentes por aplicación
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Bitácora temporal } }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: atalaya_session